Added configuration as variables for Github Action workflow #deploy

+++ terraform/instance/startup.sh
@@ -14,40 +14,44 @@ DEVICE_NAME="google-anki-data-disk"
 MOUNT_PATH="/home"
 DISK_PATH="/dev/disk/by-id/$DEVICE_NAME"
 
-echo "Checking for persistent disk at $DISK_PATH..."
-while [ ! -b "$DISK_PATH" ]; do
-  echo "Waiting for disk $DEVICE_NAME to attach..."
-  sleep 2
-done
+if mountpoint -q "$MOUNT_PATH"; then
+  echo "Disk is already mounted to $MOUNT_PATH. Skipping setup."
+else
 
-# Format the disk only if it doesn't have a filesystem yet
-if ! blkid "$DISK_PATH" > /dev/null; then
-  echo "New disk detected. Formatting with ext4..."
-  sudo mkfs.ext4 -F -E lazy_itable_init=0,lazy_journal_init=0,discard "$DISK_PATH"
+  echo "Checking for persistent disk at $DISK_PATH..."
+  while [ ! -b "$DISK_PATH" ]; do
+    echo "Waiting for disk $DEVICE_NAME to attach..."
+    sleep 2
+  done
+
+  # Format the disk only if it doesn't have a filesystem yet
+  if ! blkid "$DISK_PATH" > /dev/null; then
+    echo "New disk detected. Formatting with ext4..."
+    sudo mkfs.ext4 -F -E lazy_itable_init=0,lazy_journal_init=0,discard "$DISK_PATH"
+  fi
+
+  # Mount the disk to a temporary location first if /home is not empty
+  # (This ensures we don't lose the initial user directory created by GCP)
+  mkdir -p /mnt/tmp_home
+  mount "$DISK_PATH" /mnt/tmp_home
+
+  # If the disk is empty, move existing home data into it
+  if [ -z "$(ls -A /mnt/tmp_home)" ]; then
+    echo "Persistent disk is empty. Initializing with current /home data..."
+    cp -a /home/. /mnt/tmp_home/
+  fi
+
+  # Unmount and perform the final mount to /home
+  umount /mnt/tmp_home
+  mount "$DISK_PATH" "$MOUNT_PATH"
+
+  # Ensure it mounts automatically on reboots
+  if ! grep -qs "$MOUNT_PATH" /etc/fstab; then
+    echo "$DISK_PATH $MOUNT_PATH ext4 discard,defaults,nofail 0 2" >> /etc/fstab
+  fi
+
+  echo "Persistent disk mounted successfully to $MOUNT_PATH"
 fi
-
-# Mount the disk to a temporary location first if /home is not empty
-# (This ensures we don't lose the initial user directory created by GCP)
-mkdir -p /mnt/tmp_home
-mount "$DISK_PATH" /mnt/tmp_home
-
-# If the disk is empty, move existing home data into it
-if [ -z "$(ls -A /mnt/tmp_home)" ]; then
-  echo "Persistent disk is empty. Initializing with current /home data..."
-  cp -a /home/. /mnt/tmp_home/
-fi
-
-# Unmount and perform the final mount to /home
-umount /mnt/tmp_home
-mount "$DISK_PATH" "$MOUNT_PATH"
-
-# Ensure it mounts automatically on reboots
-if ! grep -qs "$MOUNT_PATH" /etc/fstab; then
-  echo "$DISK_PATH $MOUNT_PATH ext4 discard,defaults,nofail 0 2" >> /etc/fstab
-fi
-
-echo "Persistent disk mounted successfully to $MOUNT_PATH"
-
 # --- REST OF THE INSTALLATION SCRIPT ---
 
 TARGET_USER=$(find /home -maxdepth 1 -mindepth 1 -type d -not -name "lost+found" -printf '%f\n' | head -n 1)
@@ -86,7 +90,7 @@ if ! command -v startplasma-x11 &> /dev/null; then
 
 fi
 
-if ! command -v anki &> /dev/null; then
+if [ ! -d "$HOME_DIR/anki-launcher-25.09-linux" ]; then
   # Get Anki Installable
   echo "Installing Anki"
   wget -P "$HOME_DIR" https://github.com/ankitects/anki/releases/download/25.09/anki-launcher-25.09-linux.tar.zst
@@ -95,8 +99,8 @@ if ! command -v anki &> /dev/null; then
 
   # Add anki to autostart
   echo "Adding anki to autostart"
-  mkdir -p ~/.config/autostart
-  cat <<EOF > ~/.config/autostart/anki.desktop
+  mkdir -p "$HOME_DIR/.config/autostart"
+  cat <<EOF > "$HOME_DIR/.config/autostart/anki.desktop"
 [Desktop Entry]
 Type=Application
 Exec=anki
